- make sendEMails external (utils) - replace fromHex from libsodium-wrapper
